Segregate methods based on responsiblity
